nexinvo/backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.3
python-decouple==3.8
celery==5.3.4
redis==5.0.1
Pillow==10.2.0
reportlab==4.0.9
WeasyPrint==60.2
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
python-dateutil==2.8.2
requests==2.31.0
cryptography==42.0.2
PyJWT==2.8.0
django-storages==1.14.2
boto3==1.34.34
openpyxl==3.1.2
pandas==2.2.0
qrcode[pil]==7.4.2
python-barcode==0.15.1
lxml==5.1.0
zeep==4.2.1
oauthlib==3.2.2
requests-oauthlib==1.3.1
django-environ==0.11.2
django-redis==5.4.0
django-ratelimit==4.1.0